pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "godmode-media-library"
version = "1.0.0"
description = "GOD MODE media library organizer with metadata-first safety"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Marek Galetka" }
]
license = { text = "Proprietary" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Environment :: Console"
]

[project.optional-dependencies]
geo = [
  "geopy>=2.4.1"
]
media = [
  "Pillow>=10.0.0",
  "pillow-heif>=0.18.0"
]
people = [
  "face-recognition>=1.3.0",
  "numpy>=1.26.0",
  "Pillow>=10.0.0",
  "pillow-heif>=0.18.0",
  "scikit-learn>=1.4.0",
  "cryptography>=42.0.0"
]
watch = [
  "watchdog>=4.0.0",
]
web = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
]
full = [
  "geopy>=2.4.1",
  "face-recognition>=1.3.0",
  "numpy>=1.26.0",
  "Pillow>=10.0.0",
  "pillow-heif>=0.18.0",
  "scikit-learn>=1.4.0",
  "cryptography>=42.0.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "watchdog>=4.0.0",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "Pillow>=10.0.0",
  "httpx>=0.27.0",
  "fastapi>=0.115.0",
  "tomli>=2.0; python_version < '3.11'"
]

[project.scripts]
gml = "godmode_media_library.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"godmode_media_library.web" = ["static/**/*"]

[tool.ruff]
target-version = "py310"
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: tests that use real files and external tools",
  "requires_exiftool: tests that require ExifTool binary",
  "requires_ffprobe: tests that require ffprobe binary",
  "requires_ffmpeg: tests that require ffmpeg binary",
]
